app/lambda_src/stepfunction/ValidateResources/helper.py
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

import os
import logging
import yaml

LOG_LEVEL = os.getenv('LOG_LEVEL', 'INFO')
LOGGER = logging.getLogger()
LOGGER.setLevel(getattr(logging, LOG_LEVEL.upper(), logging.INFO))
logging.getLogger("botocore").setLevel(logging.ERROR)


def get_services_to_validate():
    """Loads yaml list of booleans for services that should be validated"""
    with open('./validate.yaml', 'rb') as file:
        return yaml.safe_load(file)
