app/lambda_src/stepfunction/ValidateResources/requirements.txt
pyyaml==5.4.1
boto3>=1.24.91